fix #9146 feat(schemas): add pop sizing schemas to package and publish to npm
