+++ .circleci/config.yml
@@ -539,7 +539,7 @@ jobs:
 
   schemas_deploy:
     docker:
-      - image: cimg/python:3.10.8 # poetry 1.2.2
+      - image: cimg/python:3.10.8-node # poetry 1.2.2
     steps:
       - checkout
       - run:
@@ -553,7 +553,7 @@ jobs:
                 circleci-agent step halt
             fi
       - run:
-          name: Create the distribution files
+          name: Create the PyPI distribution files
           command: |
             make schemas_build
       - run:
@@ -564,6 +564,19 @@ jobs:
             # For more on twine, see:
             #   https://twine.readthedocs.io/en/latest/
             make schemas_deploy_pypi
+      - run:
+          name: Create the PyPI distribution files
+          command: |
+            make schemas_build
+      - run:
+          name: NPM Authentication
+          command: echo "//registry.npmjs.org/:_authToken=${NPM_TOKEN}" > .npmrc
+      - run:
+          name: Upload to NPM
+          command: |
+            # Relies on the NPM_TOKEN environment variable configured at:
+            #   https://app.circleci.com/settings/project/github/mozilla/experimenter/environment-variables
+            make schemas_deploy_npm
 
 workflows:
   version: 2
