+++ schemas/mozilla_nimbus_schemas/jetstream/__init__.py
@@ -1,15 +1,37 @@
 from .analysis_errors import AnalysisError, AnalysisErrors
 from .metadata import ExternalConfig, Metadata, Metric, Outcome
+from .population_sizing import (
+    SampleSizes,
+    SizingByUserType,
+    SizingDetails,
+    SizingMetric,
+    SizingMetricName,
+    SizingParameters,
+    SizingRecipe,
+    SizingReleaseChannel,
+    SizingTarget,
+    SizingUserType,
+)
 from .statistics import AnalysisBasis, Statistic, Statistics
 
 __all__ = [
+    "AnalysisBasis",
     "AnalysisError",
     "AnalysisErrors",
-    "Metric",
-    "Outcome",
     "ExternalConfig",
     "Metadata",
-    "AnalysisBasis",
+    "Metric",
+    "Outcome",
+    "SampleSizes",
+    "SizingByUserType",
+    "SizingDetails",
+    "SizingMetric",
+    "SizingMetricName",
+    "SizingParameters",
+    "SizingRecipe",
+    "SizingReleaseChannel",
+    "SizingTarget",
+    "SizingUserType",
     "Statistic",
     "Statistics",
 ]
