+++ schemas/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@mozilla/nimbus-schemas",
+  "version": "2023.7.1",
+  "description": "Schemas used by Mozilla Nimbus and related projects.",
+  "main": "index.ts",
+  "scripts": {
+    "build": "openapi --input http://localhost:8000/openapi.json --output ./index.ts --client axios"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/mozilla/experimenter.git"
+  },
+  "author": "Mozilla",
+  "license": "MPL-2.0",
+  "bugs": {
+    "url": "https://github.com/mozilla/experimenter/issues"
+  },
+  "homepage": "https://github.com/mozilla/experimenter#readme",
+  "devDependencies": {
+    "openapi-typescript-codegen": "^0.25.0",
+    "typescript": "^5.1.6"
+  }
+}
